backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
pypdf
python-docx
pdf2image
PyPDF2
requests
aiohttp
sqlalchemy
python-jose[cryptography]
bcrypt